+++ front-end/src/App.js
@@ -1,10 +1,14 @@
 import * as d3 from "d3";
 import * as React from 'react';
 import Chart from "./Chart.js"
+import YearByYearChart from "./YearByYearChart.js"
 import Container from '@mui/material/Container';
 import TimeRangeSelector from "./TimeRangeSelector.js";
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
+import Select from '@mui/material/Select';
+import Input from '@mui/material/Input';
+import MenuItem from '@mui/material/MenuItem';
 
 async function fetchData(beginDate, endDate) {
   //const url = 'https://app.netatmo.net/api/getmeasure'
@@ -26,16 +30,27 @@ async function fetchData(beginDate, endDate) {
 }
 
 let initialized = false;
+const temperatureMode = 'Temperature';
+const temperatureYearByYearMode = 'Temperature year by year';
+const modes = [
+  temperatureMode,
+  temperatureYearByYearMode,
+]
 
 export default function App() {
   let [temp, setTemp] = React.useState([]);
   let [dummy, setDummy] = React.useState(1);
-  let [beginDate, setBeginDate] = React.useState(new Date(2023, 0, 1));
+  let [beginDate, setBeginDate] = React.useState(new Date(2010, 0, 1));
   let [endDate, setEndDate] = React.useState(new Date());
+  let [mode, setMode] = React.useState(temperatureYearByYearMode);
 
   React.useEffect(() => {
     fetchAndShowData();
-  }, [beginDate, endDate]);
+  }, [beginDate, endDate, mode]);
+
+  const handleModeSelectChange = (event) => {
+    setMode(event.target.value);
+  };
 
   const handleBeginDateChange = (date) => {
     setBeginDate(date);
@@ -92,16 +107,42 @@ export default function App() {
         <Typography variant="h4" component="h1" gutterBottom>
           Emil´s Weather Station @ bergetvidhandfatet
         </Typography>
-        <TimeRangeSelector
-          beginDate={beginDate}
-          endDate={endDate}
-          onBeginDateChange={handleBeginDateChange}
-          onEndDateChange={handleEndDateChange}
-        />
-        <Chart
-          data={temp}
-          dummy={dummy}
-        />
+        <Select
+          variant="standard"
+          id="mode-selector"
+          value={mode}
+          onChange={handleModeSelectChange}
+          input={<Input name="mode" id="mode-helper" />}>
+          {modes.map((mode) =>
+            <MenuItem
+              id={mode}
+              key={mode}
+              value={mode}
+            >
+              {mode}
+            </MenuItem>
+          )}
+        </Select>
+        {mode == temperatureMode && (
+          <TimeRangeSelector
+            beginDate={beginDate}
+            endDate={endDate}
+            onBeginDateChange={handleBeginDateChange}
+            onEndDateChange={handleEndDateChange}
+          />
+        )}
+        {mode == temperatureMode && (
+            <Chart
+            data={temp}
+            dummy={dummy}
+          />
+        )}
+        {mode == temperatureYearByYearMode && (
+          <YearByYearChart
+            data={temp}
+            dummy={dummy}
+          />
+        )}
       </Box>
     </Container>
   );