+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Year-by-year temperature chart. The old temperature chart is selectable through a drop-down
+
 ### Changed
 * Rename page to Emil's Weather Station @ bergetvidhandfatet
 * Change favicon to cloudy weater